requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
google-generativeai==0.3.2
langchain==0.1.5
langchain-google-genai==0.0.7
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
html2text==2020.1.16 